+++ app/controllers/home_controller.rb
@@ -10,6 +10,11 @@ class HomeController < ApplicationController
       flash.delete(:notice) if I18n.t('devise.sessions.signed_in') == flash[:notice]
       flash.delete(:alert)
 
+      if session[:incoming_heh_token].present?
+        redirect_to link_health_eheart_member_path
+        return
+      end
+
       @surveys = Survey.current.viewable.non_pediatric.limit(3)
       @answer_sessions = current_user.answer_sessions.joins(:survey).where(child_id: nil).where.not(surveys: { slug: nil }).order(:locked, "surveys.name_en", :encounter).limit(3)
       @posts = posts