+++ db/migrate/20151103142331_add_heh_columns_to_users.rb
@@ -1,0 +1,11 @@
+class AddHehColumnsToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :outgoing_heh_at, :datetime
+    add_column :users, :outgoing_heh_token, :string
+    add_column :users, :incoming_heh_at, :datetime
+    add_column :users, :incoming_heh_token, :string
+
+    add_index :users, :outgoing_heh_token, unique: true
+    add_index :users, :incoming_heh_token, unique: true
+  end
+end